backend/requirements.txt
flask
flask-cors
langchain-groq
langchain-community
langchain-classic
duckduckgo-search
requests
python-dotenv